+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 #### Improvements
 #### Dependency Upgrade
 #### New Features
+* Fix #2115: Keep tekton v1alpha1 api
 
 ### 4.9.1 (2020-04-17)
 #### Bugs